requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.7.14
cffi==1.17.1
chardet==3.0.4
click==8.2.1
cryptography==45.0.5
ecdsa==0.19.1
fastapi==0.116.1
h11==0.16.0
h2==3.2.0
hpack==3.0.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==5.2.0
idna==2.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
python-jose==3.5.0
python-multipart==0.0.20
rfc3986==1.5.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
